% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inStatement.R
\name{inStatement}
\alias{inStatement}
\title{Create an in statement}
\usage{
inStatement(vect, quoted = FALSE)
}
\arguments{
\item{vect}{a vector of values to be included in an in statment}

\item{quoted}{logical; quote the values between parentheses?}
}
\value{
A set of values surrounded by parens
}
\description{
Create an in statement
}
\examples{
c <- 1:5
inStatement(c)
}
