% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullScans.R
\name{pullScans}
\alias{pullScans}
\title{Read in assessment validation data}
\usage{
pullScans(directory)
}
\arguments{
\item{directory}{the file directory for scanned assessment documents}
}
\value{
A set of values surrounded by parens
}
\description{
This function reads in data from assessment validation surveys,
concatenates the files, and produces student IDS
}
