% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryRead.R
\name{queryRead}
\alias{queryRead}
\title{Read a query or text document in as on character value}
\usage{
queryRead(fileName)
}
\arguments{
\item{fileName}{the directory or name of a text file.}
}
\value{
A character value, hooray
}
\description{
Read a query or text document in as on character value
}
\examples{
\dontrun{queryRead('exampleQuery.txt')}
}
