% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inStatement.R
\name{inStatement}
\alias{inStatement}
\title{Create an in statement}
\usage{
inStatement(vect, quoted = FALSE)
}
\arguments{
\item{vect}{a vector of values to be included in an in statment}

\item{quoted}{logical; quote the values between parentheses?}
}
\value{
A set of values surrounded by parens
}
\description{
This funciton helps create a parenthetical character list for sql
in statements. E.g., 1:5 -> (1,2,3,4,5). This can be useful when provided
a list of something by a faculty member that needs to be used in a sql
query.
}
\examples{
c <- 1:5
inStatement(c, quoted = TRUE)
}
